// apps/frontend/src/api/generated/model/generateCommunicationRequestStyle.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Alittlebyte PoC API
 * Versioned JSON contract consumed by the React SPA.
 * OpenAPI spec version: 1.0.0
 */

export type GenerateCommunicationRequestStyle = typeof GenerateCommunicationRequestStyle[keyof typeof GenerateCommunicationRequestStyle];


export const GenerateCommunicationRequestStyle = {
  Testo_informativo: 'Testo informativo',
  Avviso_operativo: 'Avviso operativo',
  Aggiornamento_breve: 'Aggiornamento breve',
} as const;