// apps/frontend/src/api/generated/model/generateCommunicationRequestTone.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Alittlebyte PoC API
 * Versioned JSON contract consumed by the Angular SPA.
 * OpenAPI spec version: 1.0.0
 */

export type GenerateCommunicationRequestTone = typeof GenerateCommunicationRequestTone[keyof typeof GenerateCommunicationRequestTone];


export const GenerateCommunicationRequestTone = {
  Chiaro_e_diretto: 'Chiaro e diretto',
  Più_istituzionale: 'Più istituzionale',
  Più_sintetico: 'Più sintetico',
  Empatico: 'Empatico',
  Tecnico: 'Tecnico',
} as const;